pyproject.toml
[project]
name = "codebase-ingestion"
version = "0.1.0"
description = "Codebase ingestion"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "tree-sitter",
    "tree-sitter-python",
    "tree-sitter-javascript",
    "pathspec",
    "dotenv>=0.9.9",
    "fastmcp>=2.10.5",
    "helix-py>=0.2.21",
    "google-genai>=1.25.0",
    "tree-sitter-rust==0.23.2",
    "rich>=14.0.0",
]
